ch06/ex05-any/intset.go
package intset

import (
	"bytes"
	"fmt"
)

const size = 32 << (^uint(0) >> 63)

type IntSet struct {
	words []uint64
}

func NewIntSet(vals []int) *IntSet {
	intSet := &IntSet{}

	for _, v := range vals {
		intSet.Add(v)
	}

	return intSet
}

func (s *IntSet) Elems() []int {
	elems := make([]int, 0)
	for i, word := range s.words {
		if word == 0 {
			continue
		}
		for j := 0; j < size; j++ {
			if word&(1<<uint(j)) != 0 {
				elems = append(elems, i*size+j)
			}
		}
	}

	return elems
}

func (s *IntSet) Add(x int) {
	word, bit := x/size, uint(x%size)
	for word >= len(s.words) {
		s.words = append(s.words, 0)
	}
	s.words[word] |= 1 << bit
}

func (s *IntSet) String() string {
	var buf bytes.Buffer
	buf.WriteByte('{')
	for i, word := range s.words {
		if word == 0 {
			continue
		}
		for j := 0; j < size; j++ {
			if word&(1<<uint(j)) != 0 {
				if buf.Len() > len("{") {
					buf.WriteByte(' ')
				}
				fmt.Fprintf(&buf, "%d", size*i+j)
			}
		}
	}
	buf.WriteByte('}')
	return buf.String()
}
